requirements.txt
discord.py==2.4.0
Flask==3.1.0
requests==2.25.1